refactor(nvim): move nvim-tree config from after to lazy
